+++ src/diceThrower.ts
@@ -1,16 +1,12 @@
-import {
-    AmmoJSPlugin,
-    Color3,
-    Engine,
-    Mesh,
-    PBRMaterial,
-    PhysicsImpostor,
-    Ray,
-    Scene,
-    SceneLoader,
-    Vector3,
-} from "@babylonjs/core";
-
+import { Ray } from "@babylonjs/core/Culling/ray";
+import { Engine } from "@babylonjs/core/Engines/engine";
+import { SceneLoader } from "@babylonjs/core/Loading/sceneLoader";
+import { PBRMaterial } from "@babylonjs/core/Materials/PBR/pbrMaterial";
+import { Color3, Vector3 } from "@babylonjs/core/Maths/math";
+import { Mesh } from "@babylonjs/core/Meshes/mesh";
+import { PhysicsImpostor } from "@babylonjs/core/Physics/physicsImpostor";
+import { AmmoJSPlugin } from "@babylonjs/core/Physics/Plugins/ammoJSPlugin";
+import { Scene } from "@babylonjs/core/scene";
 import { Dice, DieOptions } from "./types";
 import { loadAmmo } from "./utils/ammo";
 import { getColliderFromDie, getValueFromFace, stringToDice } from "./utils/geom";